create register user endpoint

+++ backend/src/main/java/app/fitnessfinderapp/backend/accountRole/AccountRoleRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AccountRoleRepository extends JpaRepository<AccountRole, Long> {
+public interface AccountRoleRepository extends JpaRepository<AccountRole, Integer> {
   Optional<AccountRole> findByAuthority(String authority);
 }
